icmp packet binery level
to create a icmp raw socket needs to be used. it is on layer 3 and socket is layer 4. so the header needs to be crufted manually. the ip header is present in both tcp and icmp packet. tcp packets needs it's own header and icmp packets need it's own header along side with the ip header