full RAW  packet
uses SOCK_RAW and IPPROTO_RAW which means socket is raw. and the packet header has to be crufted manually. UDP TCP has built in packet system in C. but while using IPPROTO_RAW all the header and everything has to be mentioned manually. it will be manually either TCP or UDP or anything