raw socket| sending tcp with message
Sockets work on layer 4. on machine sockets can work in C without mentioning network interface. the operating system routs the connection through a network interface by it self. but raw sockets can be used to connect with a specific network interface. this case need be used with a network interface api from os.